U fix exe creation tool classcastexception when running from gradle
