+++ src/main/java/io/github/fvarrui/javapackager/gradle/GradleContext.java
@@ -12,8 +12,11 @@ import org.gradle.jvm.toolchain.JavaToolchainService;
 import org.gradle.jvm.toolchain.JavaToolchainSpec;
 
 import edu.sc.seis.launch4j.tasks.Launch4jLibraryTask;
+import io.github.fvarrui.javapackager.packagers.AbstractCreateWindowsExe;
 import io.github.fvarrui.javapackager.packagers.Context;
+import io.github.fvarrui.javapackager.packagers.CreateWindowsExeWinRun4j;
 import io.github.fvarrui.javapackager.packagers.Packager;
+import io.github.fvarrui.javapackager.packagers.WindowsPackager;
 
 /**
  * Gradle context
@@ -108,5 +111,21 @@ public class GradleContext extends Context<Logger> {
 		return super.getDefaultToolchain();
 		
 	}
+
+	@Override
+	public File createWindowsExe(WindowsPackager packager) throws Exception {
+		AbstractCreateWindowsExe createWindowsExe;
+		switch (packager.getWinConfig().getExeCreationTool()) {
+			case launch4j: createWindowsExe = new CreateWindowsExeLaunch4j(); break;
+			case winrun4j: createWindowsExe = new CreateWindowsExeWinRun4j(); break;
+			default: return null;
+		}
+		if (!createWindowsExe.skip(packager)) {
+			return createWindowsExe.apply(packager);
+		}
+		return null;
+	}
+	
+
 
 }
